remove mangaesp
