--- lib-multisrc/mangaesp/assets/i18n/messages_en.properties
@@ -1,14 +1,0 @@
-search_length_error=The query must have at least 2 characters
-comics_list_error=Comics list not found
-comic_data_error=Comic data not found
-sort_by_filter_title=Sort by
-sort_by_filter_name=Name
-sort_by_filter_views=Views
-sort_by_filter_updated=Updated
-sort_by_filter_added=Added
-status_filter_title=Status
-status_filter_all=All
-status_filter_ongoing=Ongoing
-status_filter_hiatus=Hiatus
-status_filter_dropped=Dropped
-status_filter_completed=Completed
