--- lib-multisrc/mangaesp/assets/i18n/messages_es.properties
@@ -1,14 +1,0 @@
-search_length_error=La búsqueda debe tener al menos 2 caracteres
-comics_list_error=No se pudo encontrar la lista de comics
-comic_data_error=No se pudo encontrar los datos del comic
-sort_by_filter_title=Ordenar por
-sort_by_filter_name=Nombre
-sort_by_filter_views=Vistas
-sort_by_filter_updated=Actualización
-sort_by_filter_added=Agregado
-status_filter_title=Estado
-status_filter_all=Todos
-status_filter_ongoing=En curso
-status_filter_hiatus=En pausa
-status_filter_dropped=Abandonado
-status_filter_completed=Finalizado
