+++ src/es/traduccionesmoonlight/src/eu/kanade/tachiyomi/extension/es/traduccionesmoonlight/TraduccionesMoonlightDto.kt
@@ -1,24 +1,20 @@
-package eu.kanade.tachiyomi.multisrc.mangaesp
+package eu.kanade.tachiyomi.extension.es.traduccionesmoonlight
 
 import eu.kanade.tachiyomi.source.model.SChapter
 import eu.kanade.tachiyomi.source.model.SManga
+import keiyoushi.utils.tryParse
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import java.text.SimpleDateFormat
 import java.util.Locale
 
 @Serializable
+class ResponseDto<T>(
+    val response: T,
+)
+
+@Serializable
 class TopSeriesDto(
-    val response: TopSeriesResponseDto,
-)
-
-@Serializable
-class LastUpdatesDto(
-    val response: List<SeriesDto>,
-)
-
-@Serializable
-class TopSeriesResponseDto(
     @SerialName("mensual") val topMonthly: List<List<PayloadSeriesDto>>,
     @SerialName("semanal") val topWeekly: List<List<PayloadSeriesDto>>,
     @SerialName("diario") val topDaily: List<List<PayloadSeriesDto>>,
@@ -44,9 +40,6 @@ class SeriesDto(
     val trending: TrendingDto? = null,
     @SerialName("autors") private val authors: List<AuthorDto> = emptyList(),
     private val artists: List<ArtistDto> = emptyList(),
-    @Suppress("unused") // Used in some sources
-    @SerialName("idioma")
-    val language: String? = null,
 ) {
     fun toSManga(seriesPath: String): SManga = SManga.create().apply {
         title = name
@@ -93,6 +86,8 @@ class DetailDataNameDto(
     val name: String,
 )
 
+private val dateFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", Locale.US)
+
 @Serializable
 class ChapterDto(
     @SerialName("num") private val number: Float,
@@ -105,15 +100,7 @@ class ChapterDto(
         if (!this@ChapterDto.name.isNullOrBlank()) {
             name += " - ${this@ChapterDto.name}"
         }
-        date_upload = try {
-            DATE_FORMATTER.parse(date)?.time ?: 0L
-        } catch (e: Exception) {
-            0L
-        }
+        date_upload = dateFormat.tryParse(date)
         url = "$seriesPath/$seriesSlug/$slug"
-    }
-
-    companion object {
-        private val DATE_FORMATTER by lazy { SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", Locale.US) }
     }
 }
